requirements.txt
Django==5.0.6
gunicorn
djangorestframework
psycopg2-binary

